Added local diffuse cubemap. Improved pointer asset handling.

+++ TeleportServer/GeometryEncoder.cpp
@@ -226,7 +226,7 @@ avs::Result GeometryEncoder::encodeMeshes(std::vector<avs::uid> missingUIDs)
 			put(version);
 			static const int32_t DRACO_COMPRESSED_MESH_VERSION_NUMBER = 1;
 			put(DRACO_COMPRESSED_MESH_VERSION_NUMBER);
-			uint16_t nameLength = compressedMesh->name.length();
+			uint16_t nameLength = (uint16_t)compressedMesh->name.length();
 			put(nameLength);
 			put((uint8_t *)compressedMesh->name.data(), nameLength);
 
@@ -295,7 +295,7 @@ avs::Result GeometryEncoder::encodeNodes(std::vector<avs::uid> nodeUids)
 		avs::Node *node = geometryStore->getNode(uid);
 		putPayloadType(avs::GeometryPayloadType::Node, uid);
 		// Push name length.
-		uint16_t nameLength = node->name.length();
+		uint16_t nameLength = (uint16_t)node->name.length();
 		put(nameLength);
 		// Push name.
 		put((uint8_t *)node->name.data(), nameLength);
@@ -356,10 +356,10 @@ avs::Result GeometryEncoder::encodeNodes(std::vector<avs::uid> nodeUids)
 			}
 			if (node->data_type == avs::NodeDataType::Link)
 			{
-				uint16_t urlLength = node->url.length();
+				uint16_t urlLength = (uint16_t)node->url.length();
 				put(urlLength);
 				put((uint8_t *)node->url.data(), urlLength);
-				uint16_t queryLength = node->query_url.length();
+				uint16_t queryLength = (uint16_t)node->query_url.length();
 				put(queryLength);
 				put((uint8_t *)node->query_url.data(), queryLength);
 			}
@@ -384,7 +384,7 @@ avs::Result GeometryEncoder::encodeSkeleton(avs::uid skeletonID)
 	{
 		putPayloadType(avs::GeometryPayloadType::Skeleton, skeletonID);
 		// Push name length.
-		uint16_t nameLength = skeleton->name.length();
+		uint16_t nameLength = (uint16_t)skeleton->name.length();
 		put(nameLength);
 		// Push name.
 		put((uint8_t *)skeleton->name.data(), nameLength);
@@ -430,7 +430,7 @@ avs::Result GeometryEncoder::encodeAnimation(avs::uid animationID)
 		putPayloadType(avs::GeometryPayloadType::Animation, animationID);
 
 		// Push name length.
-		uint16_t nameLength = animation->name.length();
+		uint16_t nameLength = (uint16_t)animation->name.length();
 		put(nameLength);
 		// Push name.
 		put((uint8_t *)animation->name.data(), nameLength);
@@ -575,7 +575,7 @@ avs::Result GeometryEncoder::encodeMaterials(std::vector<avs::uid> missingUIDs)
 			}
 			putPayloadType(avs::GeometryPayloadType::Material, uid);
 
-			uint16_t nameLength = material->name.length();
+			uint16_t nameLength = (uint16_t)material->name.length();
 
 			// Push name length.
 			put(nameLength);
@@ -744,7 +744,7 @@ avs::Result GeometryEncoder::encodeTexturesBackend(std::vector<avs::uid> missing
 			}
 			// Place payload type onto the buffer.
 			putPayloadType(avs::GeometryPayloadType::Texture, uid);
-			uint16_t nameLength = texture->name.length();
+			uint16_t nameLength = (uint16_t)texture->name.length();
 
 			// Push name length.
 			put(nameLength);
